requirements.txt
absl-py==0.9.0
alembic==1.4.2
altair==4.1.0
astunparse==1.6.3
attrs==19.3.0
Babel==2.8.0
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gast==0.3.3
google-api-core==1.20.0
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-cloud-language==1.3.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.17
tb-nightly==2.3.0a20200621
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020062101
tf-nightly==2.3.0.dev20200621
toolz==0.10.0
tweepy==3.8.0
urllib3==1.25.9
websocket-client==0.48.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0

# requirements for heroku
psycopg2==2.7.3.1
gunicorn==19.7.1